+++ server/routes/sentencingRoutes.ts
@@ -1,8 +1,19 @@
 import { RequestHandler } from 'express'
-import type { CourtCaseCaseOutcomeAppliedAllForm, CourtCaseOverallConvictionDateForm, SentenceLengthForm } from 'forms'
+import type {
+  CourtCaseAlternativeSentenceLengthForm,
+  CourtCaseCaseOutcomeAppliedAllForm,
+  CourtCaseOverallConvictionDateForm,
+  OffenceAlternativeSentenceLengthForm,
+  SentenceLengthForm,
+} from 'forms'
+import type { Offence } from 'models'
+import deepmerge from 'deepmerge'
 import trimForm from '../utils/trim'
 import CourtAppearanceService from '../services/courtAppearanceService'
-import { sentenceLengthToSentenceLengthForm } from '../utils/mappingUtils'
+import {
+  sentenceLengthToAlternativeSentenceLengthForm,
+  sentenceLengthToSentenceLengthForm,
+} from '../utils/mappingUtils'
 import AppearanceOutcomeService from '../services/appearanceOutcomeService'
 
 export default class SentencingRoutes {
@@ -34,7 +45,7 @@ export default class SentencingRoutes {
       overallSentenceLengthForm,
       errors: req.flash('errors') || [],
       backLink: submitToCheckAnswers
-        ? `/person/${nomsId}/${addOrEditCourtCase}/${courtCaseReference}/${addOrEditCourtAppearance}/${appearanceReference}/check-overall-answers`
+        ? `/person/${nomsId}/${addOrEditCourtCase}/${courtCaseReference}/${addOrEditCourtAppearance}/${appearanceReference}/SENTENCING/check-overall-answers`
         : `/person/${nomsId}/${addOrEditCourtCase}/${courtCaseReference}/${addOrEditCourtAppearance}/${appearanceReference}/task-list`,
     })
   }
@@ -59,6 +70,65 @@ export default class SentencingRoutes {
     }
     if (submitToCheckAnswers) {
       // TODO is this the correct check your answers?
+      return res.redirect(
+        `/person/${nomsId}/${addOrEditCourtCase}/${courtCaseReference}/${addOrEditCourtAppearance}/${appearanceReference}/SENTENCING/check-overall-answers`,
+      )
+    }
+    return res.redirect(
+      `/person/${nomsId}/${addOrEditCourtCase}/${courtCaseReference}/${addOrEditCourtAppearance}/${appearanceReference}/SENTENCING/overall-conviction-date`,
+    )
+  }
+
+  public getAlternativeSentenceLength: RequestHandler = async (req, res): Promise<void> => {
+    const { nomsId, courtCaseReference, appearanceReference, addOrEditCourtCase, addOrEditCourtAppearance } = req.params
+    const { submitToCheckAnswers } = req.query
+    let courtCaseAlternativeSentenceLengthForm = (req.flash('courtCaseAlternativeSentenceLengthForm')[0] ||
+      {}) as CourtCaseAlternativeSentenceLengthForm
+
+    if (Object.keys(courtCaseAlternativeSentenceLengthForm).length === 0) {
+      courtCaseAlternativeSentenceLengthForm =
+        sentenceLengthToAlternativeSentenceLengthForm<CourtCaseAlternativeSentenceLengthForm>(
+          this.courtAppearanceService.getOverallCustodialSentenceLength(req.session, nomsId),
+        )
+    }
+    this.courtAppearanceService.setHasOverallSentenceLengthTrue(req.session, nomsId)
+    return res.render('pages/sentencing/alternative-sentence-length', {
+      nomsId,
+      courtCaseReference,
+      appearanceReference,
+      addOrEditCourtCase,
+      addOrEditCourtAppearance,
+      submitToCheckAnswers,
+      courtCaseAlternativeSentenceLengthForm,
+      isAddOffences: this.isAddJourney(addOrEditCourtCase, addOrEditCourtAppearance),
+      errors: req.flash('errors') || [],
+      backLink: `/person/${nomsId}/${addOrEditCourtCase}/${courtCaseReference}/${addOrEditCourtAppearance}/${appearanceReference}/overall-sentence-length${submitToCheckAnswers ? '?submitToCheckAnswers=true' : ''}`,
+    })
+  }
+
+  public submitAlternativeSentenceLength: RequestHandler = async (req, res): Promise<void> => {
+    const { nomsId, courtCaseReference, appearanceReference, addOrEditCourtCase, addOrEditCourtAppearance } = req.params
+    const { submitToCheckAnswers } = req.query
+    const courtCaseAlternativeSentenceLengthForm = trimForm<CourtCaseAlternativeSentenceLengthForm>(req.body)
+    const errors = this.courtAppearanceService.setOverallAlternativeSentenceLength(
+      req.session,
+      nomsId,
+      courtCaseAlternativeSentenceLengthForm,
+    )
+    if (errors.length > 0) {
+      req.flash('errors', errors)
+      req.flash('courtCaseAlternativeSentenceLengthForm', { ...courtCaseAlternativeSentenceLengthForm })
+      return res.redirect(
+        `/person/${nomsId}/${addOrEditCourtCase}/${courtCaseReference}/${addOrEditCourtAppearance}/${appearanceReference}/SENTENCING/alternative-overall-sentence-length${submitToCheckAnswers ? '?submitToCheckAnswers=true' : ''}`,
+      )
+    }
+    if (addOrEditCourtAppearance === 'edit-court-appearance') {
+      // TODO need to test this edit journey
+      return res.redirect(
+        `/person/${nomsId}/${addOrEditCourtCase}/${courtCaseReference}/${addOrEditCourtAppearance}/${appearanceReference}/SENTENCING/details`,
+      )
+    }
+    if (submitToCheckAnswers) {
       return res.redirect(
         `/person/${nomsId}/${addOrEditCourtCase}/${courtCaseReference}/${addOrEditCourtAppearance}/${appearanceReference}/SENTENCING/check-overall-answers`,
       )
@@ -209,9 +279,9 @@ export default class SentencingRoutes {
       await this.appearanceOutcomeService.getOutcomeByUuid(appearanceOutcomeUuid, req.user.username)
     ).outcomeName
     const isAddJourney = this.isAddJourney(addOrEditCourtCase, addOrEditCourtAppearance)
-    let backLink = `/person/${nomsId}/${addOrEditCourtCase}/${courtCaseReference}/${addOrEditCourtAppearance}/${appearanceReference}/sentencing/overall-conviction-date`
+    let backLink = `/person/${nomsId}/${addOrEditCourtCase}/${courtCaseReference}/${addOrEditCourtAppearance}/${appearanceReference}/SENTENCING/overall-conviction-date`
     if (isAddJourney) {
-      backLink = `/person/${nomsId}/${addOrEditCourtCase}/${courtCaseReference}/${addOrEditCourtAppearance}/${appearanceReference}/sentencing/case-outcome-applied-sentencing`
+      backLink = `/person/${nomsId}/${addOrEditCourtCase}/${courtCaseReference}/${addOrEditCourtAppearance}/${appearanceReference}/SENTENCING/case-outcome-applied-all`
     }
     return res.render('pages/sentencing/check-overall-answers', {
       nomsId,
