+++ server/views/pages/sentencing/overall-sentence-length.njk
@@ -40,7 +40,7 @@
             }]
     }) }}
     <p class="govuk-body govuk-!-margin-top-2">
-        <a href="/person/{{ nomsId }}/{{ addOrEditCourtCase }}/{{ courtCaseReference }}/{{ addOrEditCourtAppearance }}/{{ appearanceReference }}/alternative-overall-sentence-length{{ "?submitToCheckAnswers=true" if submitToCheckAnswers }}">The
+        <a href="/person/{{ nomsId }}/{{ addOrEditCourtCase }}/{{ courtCaseReference }}/{{ addOrEditCourtAppearance }}/{{ appearanceReference }}/SENTENCING/alternative-overall-sentence-length{{ "?submitToCheckAnswers=true" if submitToCheckAnswers }}">The
             sentence length is formatted differently</a>
     </p>
 {% endset %}
